requirements.txt
anthropic
streamlit
pdfplumber
python-pptx
